refactor validation code to separate file
